fix: 좌석 SSE 연결을 페이지 단일 연결로 통합해 구역 선택 시 실시간 연결 실패 해결
백엔드가 유저당 SSE 연결을 1개만 유지해 SelectSection과 SeatSection이 각각
연결을 열면 두 번째 연결이 이벤트를 받지 못하고 끊김. BookingPage에서 연결
하나만 열고 캐시 갱신은 applySeatChange 공용 함수로 처리.
